[#3558] ConditionCorrector extraction
